requirements.txt
discord.py==2.3.2
python-dotenv==1.0.1
SQLAlchemy==2.0.35
pydantic==2.9.2